requirements.txt
diffusers==0.24.0
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
accelerate==0.25.0
opencv-python==4.8.1.78
matplotlib==3.8.2
tqdm==4.66.1
einops==0.7.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0